Move files to db module
